requirements.txt
# Core dependencies
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-chroma>=0.1.0
beautifulsoup4>=4.12.2
requests>=2.31.0
selenium>=4.11.2
webdriver-manager>=4.0.0
chromadb>=0.4.13
openai>=1.3.0
python-dotenv>=1.0.0
tiktoken>=0.5.1
colorama>=0.4.6

# Data processing
numpy>=1.24.3
pandas>=2.0.3

# UI
rich>=13.5.2
